refactor migration project workspace helpers
