+++ .sampo/changesets/issue-355-migration-project-workspace.md
@@ -1,0 +1,5 @@
+---
+npm/@wp-typia/project-tools: patch
+---
+
+Split migration project workspace state and config helpers into focused runtime modules without changing migration behavior.
